fix: fixup render
