+++ gno.land/cmd/gnoland/testdata/addpkg.txtar
@@ -7,7 +7,7 @@ gnoland start
 gnokey maketx addpkg -pkgdir $WORK -pkgpath gno.land/r/foobar/bar -gas-fee 1000000ugnot -gas-wanted 2000000 -broadcast -chainid=tendermint_test test1
 
 ## execute Render
-gnokey maketx call -pkgpath gno.land/r/foobar/bar -func RenderThanNotExist -gas-fee 1000000ugnot -gas-wanted 2000000 -args '' -broadcast -chainid=tendermint_test test1
+gnokey maketx call -pkgpath gno.land/r/foobar/bar -func Render -gas-fee 1000000ugnot -gas-wanted 2000000 -args '' -broadcast -chainid=tendermint_test test1
 
 ## compare render
 cmp stdout stdout.golden
